requirements.txt
flask
waitress
werkzeug
openai-whisper
